Added necessary parameters for AuthRequest
Removed @Header parameter from the getAuth() function as the header needs to be generated at runtime. Changed the HTTP method to POST. Added @FormUrlEncoded annotation to the method. Added a @Field parameter with key grant_type. Client credentials will be the value needed (maybe overload that function to provide a default value).