New getAuthToken Function
The new getAuthToken() method calls the API and handles response will OnResponse and OnFailure. OnResponse it will save the access token and the expiration. I'll have to instead store the DateTime.now() + expires_in to actually compare it to the time onResume is called. Implemented onResume lifecycle method that will check if we have a valid token in the session storage and if not it will get one for us.